requirements.txt
aiogram==3.2.0
aiohttp~=3.9.0
Flask==2.3.3
SQLAlchemy==1.4.49
python-dotenv==1.0.0
asyncpg==0.29.0
